grsec: Apply grsecurity-2.9.1-3.4.3-201206182054.patch
commit f641a7e5f15df0827343698a120a9e7b3d27441f
Merge: e40b896 60f9a98
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:09:29 2012 -0400

    Merge branch 'pax-test' into grsec-test

commit 60f9a988efec39eabb69231ef5846cdf92020c5e
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:09:15 2012 -0400

    Update to pax-linux-3.4.3-test14.patch

 drivers/net/wireless/rt2x00/rt2x00.h      |    2 +-
 drivers/net/wireless/rt2x00/rt2x00queue.c |    4 ++--
 2 files changed, 3 insertions(+), 3 deletions(-)

commit 2322a1a95b17f3299244c567a9442b10516ba20d
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:07:33 2012 -0400

    Update to pax-linux-3.4.2-test14.patch

 arch/parisc/include/asm/uaccess.h |    4 ++--
 arch/s390/include/asm/uaccess.h   |    4 ++--
 arch/tile/include/asm/uaccess.h   |    4 ++--
 arch/x86/include/asm/uaccess_32.h |    8 ++++----
 arch/x86/include/asm/uaccess_64.h |    8 ++++----
 tools/gcc/size_overflow_plugin.c  |    2 +-
 6 files changed, 15 insertions(+), 15 deletions(-)

Signature-tree: 4d2c9b395727bda8adb08330860a8b2f1d2a7942

+++ arch/s390/include/asm/uaccess.h
@@ -299,14 +299,14 @@ __compiletime_warning("copy_from_user() buffer size is not provably correct")
 static inline unsigned long __must_check
 copy_from_user(void *to, const void __user *from, unsigned long n)
 {
-	unsigned int sz = __compiletime_object_size(to);
+	size_t sz = __compiletime_object_size(to);
 
 	might_fault();
 
 	if ((long)n < 0)
 		return n;
 
-	if (unlikely(sz != -1 && sz < n)) {
+	if (unlikely(sz != (size_t)-1 && sz < n)) {
 		copy_from_user_overflow();
 		return n;
 	}
